HARDEN H1: green build (tests, clippy -D warnings, fmt)
Bugs the failing tests exposed:
- every shape layer saved since M6 could not be reopened: its paint was flattened next to the layer's Fill as a second 'fill' key (duplicate field). Now written as 'shape_fill'; old files are repaired on load (the second 'fill' of an object is renamed).
- rotating or flipping the canvas moved shape and text layers by one pixel against pixel layers: Permutation::mapping used the index form (h-1-y) on continuous coordinates; now h-y.
- the Line tool put the whole weight on one side of the drag; it is centred now.
- SetShape marked the union of the old and new boxes dirty (a far move redrew everything between) and changed the layer before refusing a non-finite matrix.
- text line boxes ignored the alignment offset and spanned the wrap width (parley 0.9 metrics); an empty text layer reported an ink box.
Test fixes (the code was right): rotations need ok_with_ops; crop compares the geometry only; a text crop that did not move the origin; derived images start dirty; the Type counter counts every type layer; undo restores the pre-edit cache; float tolerance; a hexagon has a bottom vertex.
Clippy: auto-fixes plus [workspace.lints.clippy] allowing large_enum_variant, type_complexity and needless_range_loop (reasons in Cargo.toml); Rng::next renamed next_unit.

+++ crates/fx-core/src/command/m8.rs
@@ -294,7 +294,7 @@ pub(super) fn red_eye(
 	let redness = |p: [f32; 4]| -> f32 { (p[0] - p[1].max(p[2])) * p[3] };
 	let red: Vec<f32> = (0..w * h).map(|i| redness(at(&tiles, x0 + (i % w) as i64, y0 + (i / w) as i64))).collect();
 	// The seed: the reddest pixel within 15 px of the click.
-	let (cx, cy) = ((px - x0) as i64, (py - y0) as i64);
+	let (cx, cy) = ((px - x0), (py - y0));
 	let mut seed = None;
 	let mut best = 0.08f32;
 	for dy in -15..=15i64 {
